Refactoring of iKartRosBridge Move bridgeModule in own header file Properly close all ports and free pointers Use ros::ok() to detect whether module should be closed, so module closes gracefully
