+++ src/iKartRosBridge/CMakeLists.txt
@@ -29,7 +29,7 @@ rosbuild_genmsg()
 set(EXECUTABLE_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/bin)
 #set the default path for built libraries to the "lib" directory
 set(LIBRARY_OUTPUT_PATH ${PROJECT_SOURCE_DIR}/lib)
-rosbuild_add_executable(iKartRosBridge main.cpp bridgeThread.cpp)
+rosbuild_add_executable(iKartRosBridge main.cpp bridgeThread.cpp bridgeThread.h bridgeModule.h)
 
 target_link_libraries(iKartRosBridge ${YARP_LIBRARIES})
